requirements.txt
psutil
pymem
pillow
tk
